+++ packages/react-ui/storybook/stories/atoms/SelectField.stories.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { Box, Grid, MenuItem, TextField } from '@mui/material';
+import { Box, Grid, MenuItem, Select, TextField } from '@mui/material';
 import makeStyles from '@mui/styles/makeStyles';
 import Typography from '../../../src/components/atoms/Typography';
 import SelectField from '../../../src/components/atoms/SelectField';
@@ -216,7 +216,14 @@ const LabelAndHelperTextTemplate = ({ label, placeholder, helperText, ...rest })
   );
 };
 
-const SizeTemplate = ({ label, placeholder, defaultValue, helperText, ...rest }) => {
+const SizeTemplate = ({
+  label,
+  placeholder,
+  defaultValue,
+  helperText,
+  size,
+  ...rest
+}) => {
   const [fixedValue, setFixedValue] = useState('Twenty');
   const [fixedValue2, setFixedValue2] = useState('Ten');
   const [fixedValue3, setFixedValue3] = useState('Thirty');
@@ -244,6 +251,7 @@ const SizeTemplate = ({ label, placeholder, defaultValue, helperText, ...rest })
             label={label}
             placeholder={placeholder}
             items={menuItems}
+            size={size}
           />
         </Grid>
         <Grid item xs={3}>
@@ -253,6 +261,7 @@ const SizeTemplate = ({ label, placeholder, defaultValue, helperText, ...rest })
             label={label}
             placeholder={placeholder}
             items={menuItems}
+            size={size}
           />
         </Grid>
         <Grid item xs={3}>
@@ -262,6 +271,7 @@ const SizeTemplate = ({ label, placeholder, defaultValue, helperText, ...rest })
             label={label}
             placeholder={placeholder}
             items={menuItems}
+            size={size}
           />
         </Grid>
       </Grid>
@@ -271,31 +281,31 @@ const SizeTemplate = ({ label, placeholder, defaultValue, helperText, ...rest })
           <Typography>Empty</Typography>
         </Grid>
         <Grid item xs={3}>
-          <TextField select {...rest} variant='filled' label={label}>
+          <Select {...rest} variant='filled' label={label} size={size}>
             {menuItems.map((option) => (
               <MenuItem key={option.label} value={option.label}>
                 {option.label}
               </MenuItem>
             ))}
-          </TextField>
+          </Select>
         </Grid>
         <Grid item xs={3}>
-          <TextField select {...rest} variant='outlined' label={label}>
+          <Select {...rest} variant='outlined' label={label} size={size}>
             {menuItems.map((option) => (
               <MenuItem key={option.label} value={option.label}>
                 {option.label}
               </MenuItem>
             ))}
-          </TextField>
+          </Select>
         </Grid>
         <Grid item xs={3}>
-          <TextField select {...rest} variant='standard' label={label}>
+          <Select {...rest} variant='standard' label={label} size={size}>
             {menuItems.map((option) => (
               <MenuItem key={option.label} value={option.label}>
                 {option.label}
               </MenuItem>
             ))}
-          </TextField>
+          </Select>
         </Grid>
       </Grid>
 
@@ -311,6 +321,9 @@ const SizeTemplate = ({ label, placeholder, defaultValue, helperText, ...rest })
             select
             value={fixedValue}
             onChange={handleChange}
+            SelectProps={{
+              size: size
+            }}
           >
             {menuItems.map((option) => (
               <MenuItem key={option.label} value={option.label}>
@@ -327,6 +340,9 @@ const SizeTemplate = ({ label, placeholder, defaultValue, helperText, ...rest })
             select
             value={fixedValue2}
             onChange={handleChange2}
+            SelectProps={{
+              size: size
+            }}
           >
             {menuItems.map((option) => (
               <MenuItem key={option.label} value={option.label}>
@@ -343,6 +359,9 @@ const SizeTemplate = ({ label, placeholder, defaultValue, helperText, ...rest })
             select
             value={fixedValue3}
             onChange={handleChange3}
+            SelectProps={{
+              size: size
+            }}
           >
             {menuItems.map((option) => (
               <MenuItem key={option.label} value={option.label}>
@@ -364,6 +383,7 @@ const SizeTemplate = ({ label, placeholder, defaultValue, helperText, ...rest })
             label={label}
             placeholder={placeholder}
             focused
+            size={size}
             items={menuItems}
           />
         </Grid>
@@ -374,6 +394,7 @@ const SizeTemplate = ({ label, placeholder, defaultValue, helperText, ...rest })
             label={label}
             placeholder={placeholder}
             focused
+            size={size}
             items={menuItems}
           />
         </Grid>
@@ -384,6 +405,7 @@ const SizeTemplate = ({ label, placeholder, defaultValue, helperText, ...rest })
             label={label}
             placeholder={placeholder}
             focused
+            size={size}
             items={menuItems}
           />
         </Grid>
@@ -400,6 +422,7 @@ const SizeTemplate = ({ label, placeholder, defaultValue, helperText, ...rest })
             label={label}
             placeholder={placeholder}
             disabled
+            size={size}
             items={menuItems}
           />
         </Grid>
@@ -410,6 +433,7 @@ const SizeTemplate = ({ label, placeholder, defaultValue, helperText, ...rest })
             label={label}
             placeholder={placeholder}
             disabled
+            size={size}
             items={menuItems}
           />
         </Grid>
@@ -420,6 +444,7 @@ const SizeTemplate = ({ label, placeholder, defaultValue, helperText, ...rest })
             label={label}
             placeholder={placeholder}
             disabled
+            size={size}
             items={menuItems}
           />
         </Grid>
@@ -437,6 +462,7 @@ const SizeTemplate = ({ label, placeholder, defaultValue, helperText, ...rest })
             placeholder={placeholder}
             helperText={helperText}
             error
+            size={size}
             items={menuItems}
           />
         </Grid>
@@ -448,6 +474,7 @@ const SizeTemplate = ({ label, placeholder, defaultValue, helperText, ...rest })
             placeholder={placeholder}
             helperText={helperText}
             error
+            size={size}
             items={menuItems}
           />
         </Grid>
@@ -459,6 +486,7 @@ const SizeTemplate = ({ label, placeholder, defaultValue, helperText, ...rest })
             placeholder={placeholder}
             helperText={helperText}
             error
+            size={size}
             items={menuItems}
           />
         </Grid>
@@ -597,7 +625,7 @@ export const LabelAndHelperText = LabelAndHelperTextTemplate.bind({});
 LabelAndHelperText.args = { ...commonArgs };
 
 export const Medium = SizeTemplate.bind({});
-Medium.args = { ...commonArgs, ...sizeArgs };
+Medium.args = { ...commonArgs, ...sizeArgs, size: 'medium' };
 Medium.argTypes = disabledControlsSizeArgTypes;
 
 export const Small = SizeTemplate.bind({});